==> notebooks/01. Data preparation.ipynb <==
# Load libraries

import pandas as pd

# Read data

df = pd.read_csv('../data/init_data.csv')
df = df.drop(columns = ['Product ID','TWF','HDF','PWF','OSF','RNF','UDI'])

# Data preparation

dummies = pd.get_dummies(df['Type'],prefix='Type')
df['Difference_of_temperature'] = df['Process temperature [K]'] - df['Air temperature [K]']
df = pd.concat([df.drop(columns = ['Type']),dummies],axis = 1)

df = df[[c for c in df if c not in ['Machine failure']] + ['Machine failure']]

df.to_csv('../data/prepared_data.csv',index = False)
