==> notebooks/03. Feature engineering.ipynb <==
# Load libraries

import pandas as pd
import numpy as np
import itertools

import warnings
warnings.filterwarnings("ignore")

# Read data

df = pd.read_csv('../data/prepared_data.csv')

# Feature engineering

col_pairs = list(itertools.combinations(df.drop(columns = ['Machine failure']).columns, 2))

for col in ['Air temperature [K]','Process temperature [K]','Rotational speed [rpm]','Torque [Nm]']:
    df[col+'_log'] = np.log(df[col])

for col in ['Air temperature [K]','Process temperature [K]','Rotational speed [rpm]','Torque [Nm]']:
    df[col+'_exp'] = np.exp(df[col])

for pair in col_pairs:
    col_name = pair[0] + '-' + pair[1]
    df[col_name] = df[pair[0]] - df[pair[1]]

for pair in col_pairs:
    col_name = pair[0] + '+' + pair[1]
    df[col_name] = df[pair[0]] + df[pair[1]]
    
for pair in col_pairs:
    col_name = pair[0] + '*' + pair[1]
    df[col_name] = df[pair[0]] * df[pair[1]]


df = df[[c for c in df if c not in ['Machine failure']] + ['Machine failure']]

df.to_csv('../data/data_fe.csv',index = False)
