==> notebooks/06. DNN creator experiments.ipynb <==
import os
import sys
sys.path.append(os.path.abspath(os.path.join(os.path.dirname('src'), '..')))

from src.trainer.nn_creater import nn_creater

trainer = nn_creater('../data/data_fe.csv',
                     '../data/feature_imporstance_WS.csv',
                     num_features_list=[5,10,30,50,60,80,100,120],
                     init_param_list=[128,256,512,1024],
                     num_layers_list=[3,5,7,10],
                     batch_sizes_list=[128,256,512,1024])

trainer.start_experiment()
